+++ Floater.pde
@@ -15,7 +15,7 @@ abstract class Floater //Do NOT modify the Floater class! Make changes in the Sp
   abstract public double getDirectionX();   
   abstract public void setDirectionY(double y);   
   abstract public double getDirectionY();   
-  abstract public void setPointDirection(int degrees);   
+  abstract public void setPointDirection(float degrees);   
   abstract public double getPointDirection(); 
 
   //Accelerates the floater in the direction it is pointing (myPointDirection)
